Modified template operations.
Template paths are registered by Template::addSearchPath instead of
using Config.  Custom template functions can also be registered by
Template::registerFunction.  Each template gets a special parameter
called 'self' that can be used to call functions, as well as call get
and send for other templates:

    echo $self->get('template', $params, $override);
    self->send('template', $params, $override);

    <?= $self->escape($value); ?>

The above is an example and requires the 'escape' function to be
registered.